refactor: Move `Hosts` to a module of `instance`
